Video 30. Adding in functionality to save the encrypted password and to delete a user and the associated customer since we do not cascade this via JPA.

+++ src/main/java/guru/springframework/services/reposervices/UserServiceRepoImpl.java
@@ -2,8 +2,10 @@ package guru.springframework.services.reposervices;
 
 import guru.springframework.domain.User;
 import guru.springframework.exceptions.NotFoundException;
+import guru.springframework.repositories.CustomerRepository;
 import guru.springframework.repositories.UserRepository;
 import guru.springframework.services.UserService;
+import guru.springframework.services.security.EncryptionService;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
@@ -16,9 +18,13 @@ import java.util.Optional;
 public class UserServiceRepoImpl implements UserService {
 
     private UserRepository userRepository;
+    private CustomerRepository customerRepository;
+    private EncryptionService encryptionService;
 
-    public UserServiceRepoImpl(UserRepository userRepository) {
+    public UserServiceRepoImpl(UserRepository userRepository, CustomerRepository customerRepository, EncryptionService encryptionService) {
         this.userRepository = userRepository;
+        this.customerRepository = customerRepository;
+        this.encryptionService = encryptionService;
     }
 
     @Override
@@ -40,11 +46,20 @@ public class UserServiceRepoImpl implements UserService {
 
     @Override
     public User saveOrUpdate(User domainObject) {
+        if(domainObject.getPassword() != null){
+            domainObject.setEncryptedPassword(encryptionService.encryptString(domainObject.getPassword()));
+        }
         return userRepository.save(domainObject);
     }
 
     @Override
     public void delete(Integer id) {
+
+        Optional<User> userOptional = userRepository.findById(id);
+        if(!userOptional.isPresent()) {
+            throw new NotFoundException("User with id " + id + " could not be found and will not be deleted!");
+        }
+        customerRepository.delete(userOptional.get().getCustomer());
         userRepository.deleteById(id);
     }
